fix(install): darwin cross-compile — uint32(st.Mode) + CI cross-compile guard (#377)
syscall.Stat_t.Mode is uint16 on darwin and uint32 on linux, so
`isDirectory(st.Mode)` compiled on the linux CI runner but broke the darwin
cross-compile in release.yml (blocking the v2.112.0 release). Cast to uint32 at
both call sites (datahome_unix.go:216,300) — identity on linux, widening on
darwin, S_IFMT bits preserved (no change to SEC-01 ownership/symlink logic).

The go-test CI job now also cross-compiles all five release targets
(linux/darwin/windows) so a GOOS-specific type break surfaces on the PR, not at
tag time — closing the held no-cross-compile-in-CI gap.

+++ cmd/install/datahome_unix.go
@@ -213,7 +213,7 @@ func createAndSecureDir(path string) error {
 		return fmt.Errorf("fstat on data-home directory %q: %w", path, err)
 	}
 
-	if !isDirectory(st.Mode) {
+	if !isDirectory(uint32(st.Mode)) {
 		return fmt.Errorf("data-home path %q opened but fstat reports it is not a directory", path)
 	}
 
@@ -297,7 +297,7 @@ func ensureAncestorsSecure(path string) error {
 		if fstatErr != nil {
 			return fmt.Errorf("fstat on ancestor directory %q: %w", dir, fstatErr)
 		}
-		if !isDirectory(st.Mode) {
+		if !isDirectory(uint32(st.Mode)) {
 			return fmt.Errorf("ancestor path %q is not a directory after creation", dir)
 		}
 	}
